python-service/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
python-multipart==0.0.6
numpy==1.26.3
scikit-learn==1.5.2
pandas==2.0.3
python-dotenv==1.0.0
websockets==12.0
python-socketio==5.11.0
aiohttp==3.9.1
